Update main.py
only get all creators from a site when that site is passed.

Instead of storing a huge list of jsons for every post then downloading. Each post is downloaded as it is read in and only chunks of 25 post jsons are stored in memory at a time.

Started working on downloading discord servers.